[005-LANDING] rename component
