+++ src/components/listing/card.vue
